+++ src/components/QuestionFlowComments/partials/Comment/Comment.jsx
@@ -19,28 +19,8 @@ constructor(props) {
     }
   }
 
-  handleOpen = () => {
-    this.setState({open: true});
-  };
-
-  handleClose = () => {
-    this.setState({open: false});
-  };
-
-  createReport = (text) => {
-    this.props.store.createReport(text)
-  }
-
-  confirmDelete() {
-    
-  }
-
-  deleteComment = () => {
-    this.props.QuestionCommentsStore.deleteComment(this.props.comment);
-  }
-
   render() {
-    const { comment, UserStore, onDelete } = this.props
+    const { comment, UserStore, onDelete, onReport } = this.props
     return (
       <div className="comment">
         {/*<Votes />*/}
@@ -71,77 +51,12 @@ constructor(props) {
             <span className="dot"> · </span>
             <span className="date">10 Sep</span>
             <span className="dot"> · </span>
-            <a className="report" onClick={this.handleOpen} >Report</a>
-
-            <ReportDialog open={this.state.open} handleClose={this.handleClose} createReport={this.createReport}/>
+            <a className="report" onClick={onReport} >Report</a>
 
           </div>
         </div>
       </div>
     );
-  }
-}
-
-class ReportDialog extends Component {
-
-  constructor(props) {
-    super(props)
-
-    this.state = {
-      reportText: ''
-    }
-  }
-
-  handleInputChange = (e) => {
-    this.setState({ reportText: e.target.value} )
-  }
-
-
-  handleSand = () => {
-    console.log(this.state.reportText)
-    const { reportText } = this.state
-    this.props.createReport(reportText)
-  }
-
-
-  render () {
-    const { open, handleClose } = this.props 
-    const { reportText } = this.state 
-
-    const actions = [
-      <FlatButton
-        label="Cancel"
-        primary={true}
-        onTouchTap={handleClose}
-      />,
-      <FlatButton
-        label="Sand"
-        primary={true}
-        onTouchTap={this.handleSand}
-      />,
-    ];
-
-    return (
-      <Dialog
-        title="Get in touch"
-        actions={actions}
-        modal={true}
-        open={open}
-      >
-      We love hearing from you and will reply to every message!
-      <br/>
-      <TextField
-        hintText="What?! WHY?! Where? When?! HOW??!?"
-        multiLine={true}
-        fullWidth
-        rowsMax={4}
-        value={reportText}
-        onChange={this.handleInputChange}
-      />
-      <br/>
-      You don’t need to say which page you’re on - it’ll tell us!
-      </Dialog>
-    )
   }
 }
