+++ src/components/QuestionFlowComments/partials/ConfirmDeleteCommentDialog/index.js
@@ -1,0 +1,2 @@
+import ConfirmDeleteCommentDialog from './ConfirmDeleteCommentDialog';
+export default ConfirmDeleteCommentDialog;
