+++ src/components/QuestionFlowComments/partials/ReportDialog/index.js
@@ -1,0 +1,2 @@
+import ReportDialog from './ReportDialog';
+export default ReportDialog;
